+++ lib/src/components/global_state.dart
@@ -1,0 +1,13 @@
+import 'package:built_collection/built_collection.dart';
+import 'package:dash_kit_core/dash_kit_core.dart';
+
+abstract class GlobalState {
+  BuiltMap<Object, OperationState> get operationsState;
+
+  T updateOperation<T extends GlobalState>(
+    Object operationKey,
+    OperationState operationState,
+  );
+
+  OperationState getOperationState(Object operationKey);
+}